Store definitions as JSON rather than plain text
Previously, definitions were stored as plain text, and were converted to expressions by parsing them. That approach was fragile and would break if the syntax were changed. I'm considering changing the syntax soon, so this change is necessary.

Definitions stored in the old format will be automatically converted to the new format.